+++ packages/widget/eslint.config.mjs
@@ -27,7 +27,8 @@ export default tseslint.config(
       "semi": ["error", "always"],
       "no-console": ["error", { allow: ["warn", "error"] }],
       "quotes": ["error", "double", { "avoidEscape": true }],
+      "@typescript-eslint/no-unsafe-function-type": "off",
     },
-    ignores: ["scripts/prepublish.cjs", "build/*"],
+    ignores: ["scripts/prepublish.cjs", "scripts/generate-chains.cjs", "build/*"],
   }
 );