+++ packages/widget/src/icons/index.tsx
@@ -35,9 +35,7 @@ type IconProps = SVGProps<SVGSVGElement> & {
   backgroundColor?: string;
 };
 
-type IconMap = {
-  [key in ICONS]: (props: IconProps) => JSX.Element | null;
-};
+type IconMap = Record<ICONS, (props: IconProps) => JSX.Element | null>;
 
 export const iconMap: IconMap = {
   [ICONS.none]: () => null,