+++ packages/widget/src/state/route.ts
@@ -1,5 +1,5 @@
 import { convertHumanReadableAmountToCryptoAmount } from "@/utils/crypto";
-import { BridgeType, ExperimentalFeature, RouteRequest, RouteResponse, SmartSwapOptions, SwapVenueRequest } from "@skip-go/client";
+import { RouteResponse, RouteConfig, RouteRequest } from "@skip-go/client";
 import { atom } from "jotai";
 import { atomWithQuery } from "jotai-tanstack-query";
 import { errorAtom } from "./errorPage";
@@ -16,6 +16,7 @@ import {
   debouncedDestinationAssetAmountValueInitializedAtom,
 } from "./swapPage";
 import { atomEffect } from "jotai-effect";
+import { WidgetRouteConfig } from "@/widget/Widget";
 
 export const initializeDebounceValuesEffect = atomEffect((get, set) => {
   const sourceAsset = get(sourceAssetAtom);
@@ -82,16 +83,7 @@ type CaughtRouteError = {
   error: unknown;
 };
 
-export type RouteConfig = {
-  experimentalFeatures?: ExperimentalFeature[];
-  allowMultiTx?: boolean;
-  allowUnsafe?: boolean;
-  bridges?: BridgeType[];
-  swapVenues?: SwapVenueRequest[];
-  smartSwapOptions?: SmartSwapOptions;
-}
-
-export const routeConfigAtom = atom<RouteConfig>({
+export const routeConfigAtom = atom<WidgetRouteConfig>({
   experimentalFeatures: ["hyperlane"],
   allowMultiTx: true,
   allowUnsafe: true,
@@ -100,6 +92,34 @@ export const routeConfigAtom = atom<RouteConfig>({
     evmSwaps: true,
   },
 });
+
+export const convertWidgetRouteConfigToClientRouteConfig = (params: WidgetRouteConfig): RouteConfig => {
+  return {
+    ...params,
+    swapVenues: params.swapVenues?.map((venue) => ({
+      ...venue,
+      chainID: venue.chainId,
+    })),
+    swapVenue: params.swapVenue && {
+      ...params.swapVenue,
+      chainID: params.swapVenue.chainId,
+    },
+  };
+}
+
+export const convertClientRouteConfigToWidgetRouteConfig = (params: RouteConfig): WidgetRouteConfig => {
+  return {
+    ...params,
+    swapVenues: params.swapVenues?.map((venue) => ({
+      ...venue,
+      chainId: venue.chainID,
+    })),
+    swapVenue: params.swapVenue && {
+      ...params.swapVenue,
+      chainId: params.swapVenue.chainID,
+    },
+  };
+}
 
 export const _skipRouteAtom = atomWithQuery((get) => {
   const skip = get(skipClient);
@@ -123,10 +143,11 @@ export const _skipRouteAtom = atomWithQuery((get) => {
         throw new Error("No route request provided");
       }
       try {
+        const _skipRouteConfig = convertWidgetRouteConfigToClientRouteConfig(skipRouteConfig);
         const response = await skip.route({
           ...params,
           smartRelay: true,
-          ...skipRouteConfig,
+          ..._skipRouteConfig,
         });
         return response;
       } catch (error) {